+++ doc/html/search/all_18.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['use_20libmagicxx_20in_20a_20cmake_20based_20project_0',['Use Libmagicxx in a CMake based Project',['../md_CONTRIBUTING.html#autotoc_md35',1,'How to Use Libmagicxx in a CMake-based Project'],['../index.html#autotoc_md47',1,'How to Use Libmagicxx in a CMake-based Project']]],
+  ['use_20libmagicxx_20in_20a_20cmake_20based_20project_0',['Use Libmagicxx in a CMake based Project',['../md_CONTRIBUTING.html#autotoc_md36',1,'How to Use Libmagicxx in a CMake-based Project'],['../index.html#autotoc_md48',1,'How to Use Libmagicxx in a CMake-based Project']]],
   ['utility_1',['utility',['../namespaceutility.html',1,'']]],
   ['utility_2ehpp_2',['utility.hpp',['../utility_8hpp.html',1,'']]]
 ];