+++ doc/html/search/all_1a.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['warning_0',['2. Warning',['../md_CODE__OF__CONDUCT.html#autotoc_md25',1,'']]]
+  ['warning_0',['2. Warning',['../md_CODE__OF__CONDUCT.html#autotoc_md26',1,'']]]
 ];