+++ doc/html/search/all_2.js
@@ -1,11 +1,13 @@
 var searchData=
 [
-  ['2_2003_2004_202024_0',['[v1.0.2] - 03-04-2024',['../md_CHANGELOG.html#autotoc_md14',1,'']]],
-  ['2_2012_2005_202024_1',['[v4.1.2] - 12-05-2024',['../md_CHANGELOG.html#autotoc_md5',1,'']]],
-  ['2_20warning_2',['2. Warning',['../md_CODE__OF__CONDUCT.html#autotoc_md25',1,'']]],
-  ['2024_3',['2024',['../md_CHANGELOG.html#autotoc_md16',1,'[v1.0.0] - 03-01-2024'],['../md_CHANGELOG.html#autotoc_md15',1,'[v1.0.1] - 11-03-2024'],['../md_CHANGELOG.html#autotoc_md14',1,'[v1.0.2] - 03-04-2024'],['../md_CHANGELOG.html#autotoc_md13',1,'[v2.0.0] - 06-04-2024'],['../md_CHANGELOG.html#autotoc_md12',1,'[v2.1.0] - 07-04-2024'],['../md_CHANGELOG.html#autotoc_md11',1,'[v3.0.0] - 14-04-2024'],['../md_CHANGELOG.html#autotoc_md10',1,'[v3.1.0] - 23-04-2024'],['../md_CHANGELOG.html#autotoc_md9',1,'[v3.1.1] - 28-04-2024'],['../md_CHANGELOG.html#autotoc_md8',1,'[v4.0.0] - 04-05-2024'],['../md_CHANGELOG.html#autotoc_md7',1,'[v4.1.0] - 05-05-2024'],['../md_CHANGELOG.html#autotoc_md6',1,'[v4.1.1] - 07-05-2024'],['../md_CHANGELOG.html#autotoc_md5',1,'[v4.1.2] - 12-05-2024'],['../md_CHANGELOG.html#autotoc_md4',1,'[v5.0.0] - 06-06-2024'],['../md_CHANGELOG.html#autotoc_md3',1,'[v5.1.0] - 23-06-2024'],['../md_CHANGELOG.html#autotoc_md2',1,'[v5.1.1] - 25-06-2024']]],
-  ['23_2004_202024_4',['[v3.1.0] - 23-04-2024',['../md_CHANGELOG.html#autotoc_md10',1,'']]],
-  ['23_2006_202024_5',['[v5.1.0] - 23-06-2024',['../md_CHANGELOG.html#autotoc_md3',1,'']]],
-  ['25_2006_202024_6',['[v5.1.1] - 25-06-2024',['../md_CHANGELOG.html#autotoc_md2',1,'']]],
-  ['28_2004_202024_7',['[v3.1.1] - 28-04-2024',['../md_CHANGELOG.html#autotoc_md9',1,'']]]
+  ['2_200_2013_2002_202025_0',['[v5.2.0] - 13-02-2025',['../md_CHANGELOG.html#autotoc_md2',1,'']]],
+  ['2_2003_2004_202024_1',['[v1.0.2] - 03-04-2024',['../md_CHANGELOG.html#autotoc_md15',1,'']]],
+  ['2_2012_2005_202024_2',['[v4.1.2] - 12-05-2024',['../md_CHANGELOG.html#autotoc_md6',1,'']]],
+  ['2_20warning_3',['2. Warning',['../md_CODE__OF__CONDUCT.html#autotoc_md26',1,'']]],
+  ['2024_4',['2024',['../md_CHANGELOG.html#autotoc_md17',1,'[v1.0.0] - 03-01-2024'],['../md_CHANGELOG.html#autotoc_md16',1,'[v1.0.1] - 11-03-2024'],['../md_CHANGELOG.html#autotoc_md15',1,'[v1.0.2] - 03-04-2024'],['../md_CHANGELOG.html#autotoc_md14',1,'[v2.0.0] - 06-04-2024'],['../md_CHANGELOG.html#autotoc_md13',1,'[v2.1.0] - 07-04-2024'],['../md_CHANGELOG.html#autotoc_md12',1,'[v3.0.0] - 14-04-2024'],['../md_CHANGELOG.html#autotoc_md11',1,'[v3.1.0] - 23-04-2024'],['../md_CHANGELOG.html#autotoc_md10',1,'[v3.1.1] - 28-04-2024'],['../md_CHANGELOG.html#autotoc_md9',1,'[v4.0.0] - 04-05-2024'],['../md_CHANGELOG.html#autotoc_md8',1,'[v4.1.0] - 05-05-2024'],['../md_CHANGELOG.html#autotoc_md7',1,'[v4.1.1] - 07-05-2024'],['../md_CHANGELOG.html#autotoc_md6',1,'[v4.1.2] - 12-05-2024'],['../md_CHANGELOG.html#autotoc_md5',1,'[v5.0.0] - 06-06-2024'],['../md_CHANGELOG.html#autotoc_md4',1,'[v5.1.0] - 23-06-2024'],['../md_CHANGELOG.html#autotoc_md3',1,'[v5.1.1] - 25-06-2024']]],
+  ['2025_5',['[v5.2.0] - 13-02-2025',['../md_CHANGELOG.html#autotoc_md2',1,'']]],
+  ['23_2004_202024_6',['[v3.1.0] - 23-04-2024',['../md_CHANGELOG.html#autotoc_md11',1,'']]],
+  ['23_2006_202024_7',['[v5.1.0] - 23-06-2024',['../md_CHANGELOG.html#autotoc_md4',1,'']]],
+  ['25_2006_202024_8',['[v5.1.1] - 25-06-2024',['../md_CHANGELOG.html#autotoc_md3',1,'']]],
+  ['28_2004_202024_9',['[v3.1.1] - 28-04-2024',['../md_CHANGELOG.html#autotoc_md10',1,'']]]
 ];