+++ doc/html/search/all_3.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['3_20temporary_20ban_0',['3. Temporary Ban',['../md_CODE__OF__CONDUCT.html#autotoc_md26',1,'']]]
+  ['3_20temporary_20ban_0',['3. Temporary Ban',['../md_CODE__OF__CONDUCT.html#autotoc_md27',1,'']]]
 ];