+++ doc/html/search/all_4.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['4_20permanent_20ban_0',['4. Permanent Ban',['../md_CODE__OF__CONDUCT.html#autotoc_md27',1,'']]]
+  ['4_20permanent_20ban_0',['4. Permanent Ban',['../md_CODE__OF__CONDUCT.html#autotoc_md28',1,'']]]
 ];