+++ doc/html/search/all_d.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['how_20to_20build_20libmagicxx_0',['How to Build Libmagicxx',['../index.html#autotoc_md46',1,'']]],
-  ['how_20to_20create_20an_20issue_1',['How to Create an Issue',['../md_CONTRIBUTING.html#autotoc_md39',1,'']]],
-  ['how_20to_20use_20libmagicxx_20in_20a_20cmake_20based_20project_2',['How to Use Libmagicxx in a CMake based Project',['../md_CONTRIBUTING.html#autotoc_md35',1,'How to Use Libmagicxx in a CMake-based Project'],['../index.html#autotoc_md47',1,'How to Use Libmagicxx in a CMake-based Project']]]
+  ['how_20to_20build_20libmagicxx_0',['How to Build Libmagicxx',['../index.html#autotoc_md47',1,'']]],
+  ['how_20to_20create_20an_20issue_1',['How to Create an Issue',['../md_CONTRIBUTING.html#autotoc_md40',1,'']]],
+  ['how_20to_20use_20libmagicxx_20in_20a_20cmake_20based_20project_2',['How to Use Libmagicxx in a CMake based Project',['../md_CONTRIBUTING.html#autotoc_md36',1,'How to Use Libmagicxx in a CMake-based Project'],['../index.html#autotoc_md48',1,'How to Use Libmagicxx in a CMake-based Project']]]
 ];